removed unwanted imports, changed folder structure
