+++ src/components/_dashboard/general-app/AppWidgets3.tsx
@@ -2,7 +2,7 @@ import { Icon } from '@iconify/react';
 import googleClassroom from '@iconify/icons-mdi/google-classroom';
 
 // material
-import { useTheme, experimentalStyled as styled } from '@material-ui/core/styles';
+import { experimentalStyled as styled } from '@material-ui/core/styles';
 import { Card, Box, Typography, CardProps } from '@material-ui/core';
 // utils
 import { fNumber } from '../../../utils/formatNumber';
@@ -33,8 +33,6 @@ interface AppWidgets3Props extends CardProps {
 }
 
 export default function AppWidgets3({ batchCount }: AppWidgets3Props) {
-  const theme = useTheme();
-
   return (
     <RootStyle>
       <Box sx={{ ml: 3, color: 'grey.800' }}>
