+++ src/components/_dashboard/general-app/AppWidgets4.tsx
@@ -2,7 +2,7 @@ import { Icon } from '@iconify/react';
 import accountChild from '@iconify/icons-mdi/account-child';
 
 // material
-import { useTheme, experimentalStyled as styled } from '@material-ui/core/styles';
+import { experimentalStyled as styled } from '@material-ui/core/styles';
 import { Card, Box, Typography, CardProps } from '@material-ui/core';
 // utils
 import { fNumber } from '../../../utils/formatNumber';
@@ -33,8 +33,6 @@ interface AppWidgets4Props extends CardProps {
 }
 
 export default function AppWidgets4({ teamCount }: AppWidgets4Props) {
-  const theme = useTheme();
-
   return (
     <RootStyle>
       <Box sx={{ ml: 3, color: 'grey.800' }}>
