--- src/components/_dashboard/user/batchList/UserList.tsx
@@ -1,248 +1,0 @@
-import { filter } from 'lodash';
-import { sentenceCase } from 'change-case';
-import { useState, useEffect } from 'react';
-import { Icon } from '@iconify/react';
-import { useSnackbar } from 'notistack';
-import closeFill from '@iconify/icons-eva/close-fill';
-// material
-import { useTheme } from '@material-ui/core/styles';
-import {
-  Card,
-  Table,
-  Stack,
-  Avatar,
-  Button,
-  Checkbox,
-  TableRow,
-  TableBody,
-  TableCell,
-  Container,
-  Typography,
-  TableContainer,
-  TablePagination
-} from '@material-ui/core';
-import MIconButton from 'components/@material-extend/MIconButton';
-// redux
-import { RootState, useDispatch, useSelector } from '../../../../redux/store';
-import { deleteUser, editUser } from '../../../../redux/slices/user';
-// routes
-import { PATH_DASHBOARD } from '../../../../routes/paths';
-// @types
-import { BatchManager, BatchMembers, UserManager } from '../../../../@types/user';
-// components
-import Label from '../../../Label';
-import Scrollbar from '../../../Scrollbar';
-import SearchNotFound from '../../../SearchNotFound';
-import { UserListHead, UserListToolbar, UserMoreMenu } from './index';
-
-// ----------------------------------------------------------------------
-
-const TABLE_HEAD = [
-  { id: 'name', label: 'Name', alignRight: false },
-  { id: 'email', label: 'Email', alignRight: false },
-  { id: 'phone', label: 'Phone', alignRight: false },
-  { id: 'role', label: 'Role', alignRight: false },
-  { id: 'approval', label: 'Approval Status', alignRight: false },
-  { id: '', label: 'Actions', alignRight: false }
-];
-
-// ----------------------------------------------------------------------
-
-type Anonymous = Record<string | number, string>;
-
-function descendingComparator(a: Anonymous, b: Anonymous, orderBy: string) {
-  if (b[orderBy] < a[orderBy]) {
-    return -1;
-  }
-  if (b[orderBy] > a[orderBy]) {
-    return 1;
-  }
-  return 0;
-}
-
-function getComparator(order: string, orderBy: string) {
-  return order === 'desc'
-    ? (a: Anonymous, b: Anonymous) => descendingComparator(a, b, orderBy)
-    : (a: Anonymous, b: Anonymous) => -descendingComparator(a, b, orderBy);
-}
-
-function applySortFilter(
-  array: BatchManager[],
-  comparator: (a: any, b: any) => number,
-  query: string
-) {
-  const stabilizedThis = array.map((el, index) => [el, index] as const);
-  stabilizedThis.sort((a, b) => {
-    const order = comparator(a[0], b[0]);
-    if (order !== 0) return order;
-    return a[1] - b[1];
-  });
-  if (query) {
-    return filter(
-      array,
-      (_batch) => _batch.batchName.toLowerCase().indexOf(query.toLowerCase()) !== -1
-    );
-  }
-  return stabilizedThis.map((el) => el[0]);
-}
-
-type UserListProps = {
-  type: string;
-  handleRequestSort: any;
-  handleSelectAllClick: any;
-  handleClick: any;
-  handleChangeRowsPerPage: any;
-  handleFilterByName: any;
-  handleDeleteUser: any;
-  page: number;
-  setPage: any;
-  order: 'asc' | 'desc';
-  selected: string[];
-  orderBy: string;
-  filterName: string;
-  rowsPerPage: number;
-  userList: BatchManager[];
-  setRefresh?: any;
-  admins: UserManager[];
-  otherUsers: UserManager[];
-};
-
-export default function UserList({
-  type,
-  handleRequestSort,
-  handleSelectAllClick,
-  handleClick,
-  handleChangeRowsPerPage,
-  handleFilterByName,
-  handleDeleteUser,
-  page,
-  setPage,
-  order,
-  selected,
-  orderBy,
-  filterName,
-  rowsPerPage,
-  userList,
-  setRefresh,
-  admins,
-  otherUsers
-}: UserListProps) {
-  const theme = useTheme();
-  const dispatch = useDispatch();
-
-  // useEffect(() => {
-  //   dispatch(getUserList());
-  // }, [dispatch]);
-
-  const emptyRows = page > 0 ? Math.max(0, (1 + page) * rowsPerPage - userList.length) : 0;
-
-  const filteredUsers = applySortFilter(userList, getComparator(order, orderBy), filterName);
-
-  const isUserNotFound = filteredUsers.length === 0;
-
-  const getMembersName = (batchMembers: BatchMembers[]) => {
-    const names = batchMembers.map((element) => element.name);
-    return names.toString();
-  };
-
-  return (
-    <Card>
-      <UserListToolbar
-        numSelected={selected.length}
-        filterName={filterName}
-        onFilterName={handleFilterByName}
-      />
-
-      <Scrollbar>
-        <TableContainer sx={{ minWidth: 800 }}>
-          <Table>
-            <UserListHead
-              order={order}
-              orderBy={orderBy}
-              headLabel={TABLE_HEAD}
-              rowCount={userList.length}
-              numSelected={selected.length}
-              onRequestSort={handleRequestSort}
-              onSelectAllClick={handleSelectAllClick}
-            />
-            <TableBody>
-              {filteredUsers
-                .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
-                .map((row) => {
-                  const { batchId, batchName, batchOwner, batchType, batchMembers } = row;
-                  const isItemSelected = selected.indexOf(batchName) !== -1;
-
-                  return (
-                    <TableRow
-                      hover
-                      key={batchId}
-                      tabIndex={-1}
-                      role="checkbox"
-                      selected={isItemSelected}
-                      aria-checked={isItemSelected}
-                    >
-                      <TableCell align="left">{sentenceCase(batchName)}</TableCell>
-                      <TableCell align="left">{batchOwner}</TableCell>
-                      <TableCell align="left">
-                        <Label
-                          variant={theme.palette.mode === 'light' ? 'ghost' : 'filled'}
-                          color={(batchType === 'ninja' && 'success') || 'info'}
-                        >
-                          {sentenceCase(batchType)}
-                        </Label>
-                      </TableCell>
-                      <TableCell align="left">{getMembersName(batchMembers)}</TableCell>
-                      {/* <TableCell align="left">{experience}</TableCell>
-                      <TableCell align="left">
-                        <Label
-                          variant={theme.palette.mode === 'light' ? 'ghost' : 'filled'}
-                          color={(approval === 'approved' && 'success') || 'error'}
-                        >
-                          {sentenceCase(approval)}
-                        </Label>
-                      </TableCell> */}
-
-                      <TableCell align="right">
-                        <UserMoreMenu
-                          setRefresh={setRefresh}
-                          currentBatch={row}
-                          onDelete={() => handleDeleteUser(batchId)}
-                          userName={batchName}
-                          admins={admins}
-                          otherUsers={otherUsers}
-                        />
-                      </TableCell>
-                    </TableRow>
-                  );
-                })}
-              {emptyRows > 0 && (
-                <TableRow style={{ height: 53 * emptyRows }}>
-                  <TableCell colSpan={6} />
-                </TableRow>
-              )}
-            </TableBody>
-            {isUserNotFound && (
-              <TableBody>
-                <TableRow>
-                  <TableCell align="center" colSpan={6} sx={{ py: 3 }}>
-                    <SearchNotFound searchQuery={filterName} />
-                  </TableCell>
-                </TableRow>
-              </TableBody>
-            )}
-          </Table>
-        </TableContainer>
-      </Scrollbar>
-
-      <TablePagination
-        rowsPerPageOptions={[5, 10, 25]}
-        component="div"
-        count={userList.length}
-        rowsPerPage={rowsPerPage}
-        page={page}
-        onPageChange={(e, page) => setPage(page)}
-        onRowsPerPageChange={(e) => handleChangeRowsPerPage}
-      />
-    </Card>
-  );
-}
