--- src/components/_dashboard/user/batchList/index.ts
@@ -1,4 +1,0 @@
-export { default as UserListHead } from './UserListHead';
-export { default as UserListToolbar } from './UserListToolbar';
-export { default as UserMoreMenu } from './UserMoreMenu';
-export { default as UserList } from './UserList';
